Move getVisibleReminders to more generic location
