+++ src/components/info/reminders.tsx
@@ -4,7 +4,7 @@ import { without } from 'lodash'
 import { visibility, selectors } from 'ducks'
 import { componentWithSize } from 'utils/mapSizesToProps'
 import { processReminders } from 'utils/processReminders'
-import { getVisibleReminders } from 'utils/pdf/generate/getVisibleReminders'
+import { getVisibleReminders } from 'utils/reminderUtils'
 import { titleCase } from 'utils/textUtils'
 import { Reminder } from 'components/info/reminder'
 import { IArmy, TAllyArmies, ICurrentArmy } from 'types/army'
